Alzdata {metaUnion}	R Documentation

Trimmed microarray dataset for Alzheimer's disease in 6 studies

Usage

	data(Alzdata)

Format

List of 6 studies: gse5281, gse48350, gse15222, gse33000, gse44772, gse36980, which are a subset of microarray datasets downloaded from GEO (http://www.ncbi.nlm.nih.gov/geo/) .

Each list of studies contains two data frames as the following:

expression.mat A data.frame, each object/row represents a probeset and each variable/column represents a sample. The two additional columns appended at the end indicate the gene Entrez ID (this column must be named as “Entrez.Gene”) and then the gene symbol.

sample.info A data.frame with 1 object of 2 variables, numbers of case and control sample.

Source

Datasets are trimmed from original studies, including only probesets for the top 30 differential expressed genes (DEGs) identified in our study (Li X et al. 2015) and the 50-150 randomly selected probesets from the original dataset.

Examples

	data(Alzdata)

[Package metaUnion version 1.0 Index]