inputExpCheck {metaUnion}	R Documentation

Check input data format and filtering

Description

Auto-run in function metaAnalysis(). Checks if the information in the input data list matches, and filters out probes without annotations or with multiple Entrez ID annotations. We use Entrez ID as the gene ID.

Usage

	inputExpCheck(inputList)

Arguments

inputList
A list object containing lists for respective studies. Each list of a particular study contains an expression matrix and a sample information vector. The expression matrix is constructed with each row representing a gene (probeset) and each column representing a sample. The order of the columns should be [case samples][control samples][two additional columns]. The two additional columns appended at the end indicate the gene Entrez ID (this column must be named as “Entrez.Gene”) and then the gene symbol. The sample information vector contains two elements, the first one represents the number of case samples and the second one represents the number of control samples

Value

A filtered data list with complete gene annotations (Entrez gene ID) and correct format for ongoing meta-analysis. The following probesets are excluded from the list: those which are expressed in less than 2 samples in each group; those which are unannotated or annotated in the wrong format.

Examples

	inputExpCheck(inputList)

[Package metaUnion version 1.0 Index]