metaInfoStat {metaUnion}	R Documentation

Generate summary table for input data

Description

This function is to summarize the input data list and calculate the number of genes found in all studies.

Usage

	metaInfoStat(inputList, metaResult, filterData=TRUE)

Arguments

inputList
A list object containing lists for respective studies. Each list of a particular study contains an expression matrix and a sample information vector. The expression values should have been log 2 transformed. The expression matrix is constructed with each row represents a gene (feature) and each column represents a sample. The order of the column should be [case samples][control samples][two additional columns]. The two additional columns appended at the end indicate the gene Entrez ID (this column must be named as “Entrez.Gene”) and then the gene symbol. The sample information vector contains two elements, the first one represents the number of case samples and the second one represents the number of control samples.

metaResult
A meta-analysis result list generated by function metaAnalysis().

filterData
Specifying whether to filter unannotated or multi-annotated probes or not, default TRUE (do filtering).

Value

A list containing two matrices. One matrix summarizes the general information about the input data list, including number of genes, and number of case and control samples in each study. The other matrix displays the numbers of found in all studies (n) and to how many are found in n-1, n-2 studies etc.

Examples

	metaEffectInfoStat(inputList, metaResult)

[Package metaUnion version 1.0 Index]