metaUnion-package {metaUnion}	R Documentation

Meta-analysis for unified Microarrays

Description

Meta-analysis using this package will include in the calculation all genes in the input studies. It does not require the same gene dimension for every study and will take into account the cases where genes are missing from some datasets. It also gives additional useful information in the output matrix including Bonferroni corrected p-value and effect direction.

Details

The DESCRIPTION file:

Package: 	 metaUnion

Type: 	 Package

Title: 	 Meta-analysis for unified Microarrays

Version: 	 1.0

Date: 	 2015-09-22

Author: 	 Jintao Long, Xinzhong Li

Maintainer: 	 Jintao Long<jintao.long@plymouth.ac.uk>

Description: 	 Meta-analysis using this package will include in the calculation all genes in the input studies. It does not require the same gene dimension for every study and will take into account the cases where genes are missing from some datasets. It also gives additional useful information in the output matrix including Bonferroni corrected p-value and effect direction.

License:

Index of help topics:

Alzdata                 Trimmed microarray dataset for Alzheimer's
                        disease in 6 studies
inputExpCheck           Check input data format and filtering
metaAnalysis            universal meta-analysis
metaInfoStat            Generate summary table for input data
metaUnion-package       Meta-analysis for unified Microarrays
metaZscoreQQplot        Q-Q plot of meta-analysis z score
sortMetaStat            Extract top ranking genes with certain
                        measurement

{metaAnalysis

Author(s)

Jintao Long, Xinzhong Li

Maintainer: Jintao Long<jintao.long@plymouth.ac.uk>

References

Marot G., et al. Moderated effect size and P-value combinations for microarray meta-analysis. Bioinformatics. 2009 Oct 15;25(20):2692-9.

Smyth, G. K. Linear models and empirical Bayes methods for assessing differential expression in microarray experiments. Stat Appl Genet Mol Biol. 2004;3:Article3

Li X., et al. Integrated genomic approaches identify major pathways and upstream regulators in late onset Alzheimer’s disease. Sci Rep. 2015 Jul 23;5:12393.

[Package metaUnion version 1.0 Index]